cannot remove realestate, some house for sale

+++ houserush/realestate/src/main/java/org/apache/servicecomb/samples/practise/houserush/realestate/service/RealestateServiceImpl.java
@@ -67,6 +67,10 @@ public class RealestateServiceImpl implements RealestateService {
 
   @Override
   public void removeRealestate(Integer id) {
+    List<String> houseSate =houseDao.findByRealestateHouseState(id);
+    if(houseSate != null && houseSate.size()>0){
+      throw new DataRetrievalFailureException("cannot remove realestate, some house for sale");
+    }
     realestateDao.delete(id);
   }
